requirements.txt
loguru
Pillow
opencv-python
numpy